Remove Bulk Similarity and BitVector-based APIs
This PR simplifies our similarity API to only support cross Tanimoto / Cosine on GPU resident data.

The other APIs were inefficient and cluttered the codebase. The bulk case is handled as a subcase of cross with dim 1 for one of the input arrays. The cross APIs with BitVector inputs can be handled in user-space, added examples in docs on how to deal with that.

For the functionality we do keep, improved coverage of unit tests.