Remove separate dispatch for per molecule kernel sizes. Now choose one kernel based on max size. (#59)
Sending off kernel32 then kernel64 was a pessimization. We lose a tiny bit of per-kernel perf from this but we need batches to send off at once, not sequentially, and we can't afford the communication overhead of instantiating more concurrent streams.

Also fixed a leftover debugging statement that was causing the MMFF benchmark to run slow by calculating RDKit energies when not necessary. Did not affect result reporting, just did extraneous work in between benchmarks.